Extract shared test HTTP server helper
