+++ server/adminAuth.test.ts
@@ -1,6 +1,7 @@
 import http from "node:http";
 import express from "express";
 import { afterEach, describe, expect, it, vi } from "vitest";
+import { bindTestHttpServer } from "./testHttpServer";
 
 const { safeLogMock } = vi.hoisted(() => ({
   safeLogMock: vi.fn(),
@@ -38,26 +39,11 @@ async function startServer() {
   );
 
   const server = http.createServer(app);
-  await new Promise<void>(resolve => server.listen(0, "127.0.0.1", resolve));
-  const address = server.address();
-
-  if (!address || typeof address === "string") {
-    server.close();
-    throw new Error("Unable to get test server address");
-  }
+  const boundServer = await bindTestHttpServer(server);
 
   return {
-    baseUrl: `http://127.0.0.1:${address.port}`,
-    close: () =>
-      new Promise<void>((resolve, reject) => {
-        server.close(error => {
-          if (error) {
-            reject(error);
-            return;
-          }
-          resolve();
-        });
-      }),
+    baseUrl: boundServer.baseUrl,
+    close: boundServer.close,
   };
 }
